Add common module
