+++ app/build.gradle.kts
@@ -53,6 +53,7 @@ dependencies {
     implementation(project(":token-images"))
     implementation(project(":text-drawable"))
     implementation(project(":token-data"))
+    implementation(project(":common"))
     implementation("androidx.appcompat:appcompat:${Versions.APP_COMPAT}")
     implementation("com.google.android.material:material:${Versions.MATERIAL}")
