+++ settings.gradle.kts
@@ -2,5 +2,6 @@ include(":token-images")
 include(":app")
 include(":text-drawable")
 include(":token-data")
+include(":common")
