+++ token-data/src/main/java/org/fedorahosted/freeotp/data/OtpTokenDatabase.kt
@@ -2,8 +2,6 @@ package org.fedorahosted.freeotp.data
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import org.fedorahosted.freeotp.data.encryption.EncryptDecrypt
-import javax.inject.Inject
 
 @Database(entities = [OtpToken::class], version = 1, exportSchema = false)
 abstract class OtpTokenDatabase : RoomDatabase() {
